requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
boto3==1.35.43
botocore==1.35.43
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
fastapi==0.115.2
geopandas==1.0.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pip==24.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.0
PyNaCl==1.5.0
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
pytz==2024.2
s3transfer==0.10.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sshtunnel==0.4.0
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
